(fact-79) Document builtin facts in a json schema
Prior to this commit there wasn't validatable documentation
for the types of built-in facts.  This schema matches against
the type of built-in facts (at least on the platforms I tested:
osx, precise, fedora18).

This initial schema only uses primitive types (string, integer,
boolean) but could be enhanced with enumerations (e.g. for kernel,
which only assumes a limited set of values) or tighter regexes
(e.g. for network addresses that follow a fixed pattern).