(fact-79) Add facter_json_output_validate.rb
Prior to this commit there was no validation of facter output
against a schema.  This commit adds an acceptance test to validate
facter output against the recently added schema.

Note that the acceptance test is currently confined off of a
number of platforms, notably those which come with ruby 1.8.7,
since that doesn't include the json gem. We should remove those
confines as acceptance jobs are enhanced to add the gem to those
platforms.

This also adds a quicky validate.rb script for dev use.